requirements.txt
discord.py==2.0.0
requests==2.31.0
fuzzywuzzy[speedup]==0.18.0